Added php require and removed the menu separately
Removed the menu code for each .php page and created a menu.php. By
this the menu is created in a separate page and the website .php sites
are connected to the menu.php. This is easily changed, if another page
needs to be added or a link is not working.